% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2}
\alias{bowtie2}
\title{bowtie2}
\usage{
bowtie2(
  design,
  outdir = file.path(tempdir(), "EpiPrepare_results", "bowtie2"),
  conda_env = "epiprepare",
  ref_file = NULL,
  threads = 8,
  verbose = TRUE
)
}
\description{
Run bowtie2
}
