% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Run FASTQC}
\usage{
fastqc(
  files,
  outdir = file.path(tempdir(), "EpiPrepare_results", "fastqc"),
  dir = file.path(tempdir(), "fastqc_temp"),
  format = "fastq",
  threads = 1,
  background = FALSE,
  force_new = FALSE,
  return_all = FALSE,
  conda_env = "epiprepare",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{files}{Paths to files of the type specified in \code{format}.}

\item{outdir}{Create all output files in the specified output directory.
Please note that this directory must exist as the program
will not create it.  If this option is not set then the
output file for each sequence file is created in the same
directory as the sequence file which was processed.}

\item{dir}{Selects a directory to be used for temporary files written when
generating report images. Defaults to system temp directory if
not specified.}

\item{format}{Bypasses the normal sequence file format detection and
forces the program to use the specified format.  Valid
formats are bam,sam,bam_mapped,sam_mapped and fastq.}

\item{threads}{Specifies the number of files which can be processed
simultaneously.  Each thread will be allocated 250MB of
memory so you shouldn't run more threads than your
available memory will cope with, and not more than
6 threads on a 32 bit machine.}

\item{force_new}{If FASTQ results for these files already exist,
overwrite them (default: \code{FALSE}).}

\item{return_all}{Return all FASTQ results located in the \code{outdir}, 
even if they are not from one of the \code{files} in this run
 (default: \code{FALSE}).}

\item{verbose}{Print messages.}

\item{...}{Additional string arguments to be passed to \code{fastqc}.}

\item{ex}{Path to \code{fastqc} software executable.}
}
\description{
Run FASTQC on fastq files.
}
\examples{

}
