% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igenomes_search.R
\name{igenomes_search}
\alias{igenomes_search}
\title{iGenomes search}
\usage{
igenomes_search(
  source = "Ensembl",
  species = "Homo_sapiens",
  build = c("hg19", "grch37"),
  type = NULL,
  tool = "bowtie2",
  name = NULL,
  extension = "fa",
  verbose = TRUE
)
}
\description{
Search iGenomes
}
