% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_narrowPeak.R
\name{import_narrowPeak}
\alias{import_narrowPeak}
\title{Import narrow peaks}
\usage{
import_narrowPeak(con)
}
\description{
@source \href{https://support.bioconductor.org/p/80535/#80537}{
Solution from Bioconductor forum}
}
