% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_file.R
\name{make_design_files}
\alias{make_design_files}
\title{Make design files}
\usage{
make_design_files(
  root,
  pattern = "*.fq.gz",
  split_batches = FALSE,
  omit_processed = TRUE,
  force_new = TRUE,
  method = c("novogene"),
  save_dir = tempdir(),
  return_table = TRUE
)
}
\description{
Makes design files.
}
