% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimgalore.R
\name{trimgalore}
\alias{trimgalore}
\title{TrimGalore}
\source{
{https://github.com/FelixKrueger/TrimGalore/issues/53}{
Issues with Conda}
}
\usage{
trimgalore(
  design,
  outdir = file.path(tempdir(), "EpiPrepare_results", "trimgalore"),
  conda_env = "epiprepare",
  threads = 8,
  verbose = TRUE
)
}
\description{
Run TrimGalore
}
